--- app/src/main/java/com/example/baostore/activities/activity_register.java
@@ -1,16 +1,0 @@
-package com.example.baostore.activities;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.example.baostore.R;
-
-public class activity_register extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
-    }
-
-}
