+++ app/src/main/res/layout/activity_login.xml
@@ -72,6 +72,7 @@
                         android:src="@drawable/ic_user_1" />
 
                     <EditText
+                        android:id="@+id/edEmail_login"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"
                         android:background="#fff"
@@ -102,6 +103,7 @@
                         android:src="@drawable/ic_lock" />
 
                     <EditText
+                        android:id="@+id/edPassword_login"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"
                         android:background="#fff"
